+++ demo/ImportLines.sql
@@ -1,0 +1,5 @@
+# Import service lines
+importdbf demo_services_header;
+importdbf demo_services_links;
+importdbf demo_services_stops;
+